Refactor and make heuristics dynamic
