+++ scripts/src/heuristics/HeuristicRegistry.py
@@ -1,0 +1,14 @@
+
+from typing import Type, TypedDict
+
+from pydantic import BaseModel
+
+from .NamedHeuristic import NamedHeuristic
+
+HeuristicRegistry: dict[str, NamedHeuristic] = {}
+
+def register_heuristic(heuristic: Type[NamedHeuristic]):
+	HeuristicRegistry[heuristic.__name__] = heuristic
+
+def get_heuristic(name):
+	return HeuristicRegistry[name]
